tutorial/requirements.txt
opencv-python
opencv-contrib-python
matplotlib
numpy
torch
torchvision
pillow
tqdm